refactor: extract lane bulk handlers into useLaneTaskActions
